# Avoid dynamic memory allocation on read path