@pdillinger 2022-08-23
I think the only way to reinstate a change like this would be to have Cache return whether it has a configured secondary cache, save that bool, and take a fast path (basic Lookup) if not configured.

(For async Lookup, CreateCallback has to be usable after the Lookup call returns.)